eric-cnam-onboarding-server/src/main/java/com/ericsson/oss/management/onboarding/utils/YAMLParseUtils.java
/*******************************************************************************
 * COPYRIGHT Ericsson 2023
 *
 *
 *
 * The copyright to the computer program(s) herein is the property of
 *
 * Ericsson Inc. The programs may be used and/or copied only with written
 *
 * permission from Ericsson Inc. or in accordance with the terms and
 *
 * conditions stipulated in the agreement/contract under which the
 *
 * program(s) have been supplied.
 ******************************************************************************/
package com.ericsson.oss.management.onboarding.utils;

import java.io.InputStream;

import org.yaml.snakeyaml.LoaderOptions;
import org.yaml.snakeyaml.Yaml;

public final class YAMLParseUtils {
    private YAMLParseUtils(){}

    public static Iterable<Object> getYamlContent(InputStream inputStream) {
        var loaderOptions = new LoaderOptions();
        loaderOptions.setAllowDuplicateKeys(false);
        final var yaml = new Yaml(loaderOptions);
        return yaml.loadAll(inputStream);
    }

}
